Refactored __main__ into a submodule
