+++ corebehrt/azure/main/__init__.py
@@ -1,0 +1,3 @@
+from corebehrt.azure.main import environment, job, pipeline, test
+
+__all__ = [environment, job, pipeline, test]
